[Security Solution] render prevalence details new document flyout in Security Solution and Discover
