+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/left/components/test_ids.ts
@@ -19,38 +19,6 @@ export const GRAPH_VISUALIZATION_TEST_ID = `${PREFIX}GraphVisualization` as cons
 
 /* Insights tab */
 
-/* Prevalence */
-
-const PREVALENCE_DETAILS_TEST_ID = `${PREFIX}PrevalenceDetails` as const;
-export const PREVALENCE_DETAILS_DATE_PICKER_TEST_ID =
-  `${PREVALENCE_DETAILS_TEST_ID}DatePicker` as const;
-export const PREVALENCE_DETAILS_UPSELL_TEST_ID = `${PREVALENCE_DETAILS_TEST_ID}Upsell` as const;
-export const PREVALENCE_DETAILS_TABLE_TEST_ID = `${PREVALENCE_DETAILS_TEST_ID}Table` as const;
-export const PREVALENCE_DETAILS_TABLE_FIELD_CELL_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}FieldCell` as const;
-export const PREVALENCE_DETAILS_TABLE_VALUE_CELL_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}ValueCell` as const;
-export const PREVALENCE_DETAILS_TABLE_PREVIEW_LINK_CELL_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}PreviewLinkCell` as const;
-export const PREVALENCE_DETAILS_TABLE_ALERT_COUNT_CELL_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}AlertCountCell` as const;
-export const PREVALENCE_DETAILS_TABLE_DOC_COUNT_CELL_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}DocCountCell` as const;
-export const PREVALENCE_DETAILS_TABLE_INVESTIGATE_IN_TIMELINE_BUTTON_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}InvestigateInTimelineButton` as const;
-export const PREVALENCE_DETAILS_TABLE_COUNT_TEXT_BUTTON_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}CountTextButton` as const;
-export const PREVALENCE_DETAILS_TABLE_HOST_PREVALENCE_CELL_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}HostPrevalenceCell` as const;
-export const PREVALENCE_DETAILS_TABLE_USER_PREVALENCE_CELL_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}UserPrevalenceCell` as const;
-export const PREVALENCE_DETAILS_TABLE_UPSELL_CELL_TEST_ID =
-  `${PREVALENCE_DETAILS_TABLE_TEST_ID}UpsellCell` as const;
-export const PREVALENCE_DETAILS_COLD_FROZEN_TIER_CALLOUT_TEST_ID =
-  `${PREVALENCE_DETAILS_TEST_ID}ColdFrozenTierCallout` as const;
-export const PREVALENCE_DETAILS_COLD_FROZEN_TIER_CALLOUT_DISMISS_BUTTON_TEST_ID =
-  `${PREVALENCE_DETAILS_TEST_ID}ColdFrozenTierCalloutDismissButton` as const;
-
 /* Entities */
 
 export const ENTITIES_DETAILS_TEST_ID = `${PREFIX}EntitiesDetails` as const;
