+++ x-pack/solutions/security/plugins/security_solution/public/flyout/document_details/right/components/insights_section.test.tsx
@@ -24,7 +24,7 @@ import { useFirstLastSeen } from '../../../../common/containers/use_first_last_s
 import { useObservedUserDetails } from '../../../../explore/users/containers/users/observed_details';
 import { useHostDetails } from '../../../../explore/hosts/containers/hosts/details';
 import { useFetchThreatIntelligence } from '../../../../flyout_v2/document/hooks/use_fetch_threat_intelligence';
-import { usePrevalence } from '../../../../flyout_v2/document/hooks/use_prevalence';
+import { usePrevalence } from '../../../../flyout_v2/prevalence/hooks/use_prevalence';
 import { mockGetFieldsData } from '../../shared/mocks/mock_get_fields_data';
 import { mockDataFormattedForFieldBrowser } from '../../shared/mocks/mock_data_formatted_for_field_browser';
 import { mockContextValue } from '../../shared/mocks/mock_context';
@@ -111,7 +111,7 @@ const mockUseHostDetails = useHostDetails as jest.Mock;
 jest.mock('../../../../explore/hosts/containers/hosts/details');
 
 jest.mock('../../../../flyout_v2/document/hooks/use_fetch_threat_intelligence');
-jest.mock('../../../../flyout_v2/document/hooks/use_prevalence');
+jest.mock('../../../../flyout_v2/prevalence/hooks/use_prevalence');
 jest.mock('../../../../flyout_v2/document/hooks/use_show_related_alerts_by_ancestry');
 jest.mock('../../../../flyout_v2/document/hooks/use_show_related_alerts_by_same_source_event');
 jest.mock('../../../../flyout_v2/document/hooks/use_show_related_alerts_by_session');
