+++ x-pack/solutions/security/plugins/security_solution/public/flyout_v2/document/components/insights_section.tsx
@@ -9,7 +9,10 @@ import { type DataTableRecord, getFieldValue } from '@kbn/discover-utils';
 import { i18n } from '@kbn/i18n';
 import React, { memo, useCallback, useMemo } from 'react';
 import { EVENT_KIND } from '@kbn/rule-data-utils';
+import { useHistory } from 'react-router-dom';
+import { useStore } from 'react-redux';
 import { EventKind } from '../constants/event_kinds';
+import { getColumns } from '../../prevalence/utils/get_columns';
 import { useRuleWithFallback } from '../../../detection_engine/rule_management/logic/use_rule_with_fallback';
 import { FLYOUT_STORAGE_KEYS } from '../constants/local_storage';
 import { PREFIX } from '../../../flyout/shared/test_ids';
@@ -18,6 +21,11 @@ import { useExpandSection } from '../../shared/hooks/use_expand_section';
 import { ThreatIntelligenceOverview } from './threat_intelligence_overview';
 import { CorrelationsOverview } from './correlations_overview';
 import { PrevalenceOverview } from './prevalence_overview';
+import { PrevalenceDetails } from '../../prevalence/prevalence';
+import { cellActionRenderer } from '../../shared/components/cell_actions';
+import { useKibana } from '../../../common/lib/kibana';
+import { flyoutProviders } from '../../shared/components/flyout_provider';
+import { useIsInSecurityApp } from '../../../common/hooks/is_in_security_app';
 
 export const INSIGHTS_SECTION_TEST_ID = `${PREFIX}InsightsSection` as const;
 
@@ -42,6 +50,12 @@ export interface InsightsSectionProps {
  * Content to be added soon.
  */
 export const InsightsSection = memo(({ hit }: InsightsSectionProps) => {
+  const { services } = useKibana();
+  const { overlays } = services;
+  const store = useStore();
+  const history = useHistory();
+  const isInSecurityApp = useIsInSecurityApp();
+
   const expanded = useExpandSection({
     storageKey: FLYOUT_STORAGE_KEYS.OVERVIEW_TAB_EXPANDED_SECTIONS,
     title: LOCAL_STORAGE_SECTION_KEY,
@@ -66,8 +80,30 @@ export const InsightsSection = memo(({ hit }: InsightsSectionProps) => {
   );
 
   const onShowThreatIntelligenceDetails = useCallback(() => {}, []);
-  const onShowPrevalenceDetails = useCallback(() => {}, []);
   const onShowCorrelationsDetails = useCallback(() => {}, []);
+  const onShowPrevalenceDetails = useCallback(() => {
+    overlays.openSystemFlyout(
+      flyoutProviders({
+        services,
+        store,
+        history,
+        children: (
+          <PrevalenceDetails
+            hit={hit}
+            investigationFields={investigationFields}
+            scopeId={''}
+            columns={getColumns(cellActionRenderer, isInSecurityApp, '')}
+          />
+        ),
+      }),
+      {
+        ownFocus: false,
+        resizable: true,
+        size: 'm',
+        type: 'overlay',
+      }
+    );
+  }, [history, hit, investigationFields, isInSecurityApp, overlays, services, store]);
 
   return (
     <ExpandableSection
